requirements.txt
# Governance Validator Dependencies
zod>=3.0.0
pyyaml>=6.0
